hack: Lazy-Load GRPC
