+++ dev/src/util.ts
@@ -17,23 +17,6 @@
 import {DocumentData} from '@google-cloud/firestore';
 
 import {randomBytes} from 'crypto';
-import {
-  CallSettings,
-  ClientConfig,
-  constructSettings,
-  createDefaultBackoffSettings,
-  GoogleError,
-  Status,
-} from 'google-gax';
-import {BackoffSettings} from 'google-gax/build/src/gax';
-import * as gapicConfig from './v1/firestore_client_config.json';
-
-const serviceConfig = constructSettings(
-  'google.firestore.v1.Firestore',
-  gapicConfig as ClientConfig,
-  {},
-  Status
-) as {[k: string]: CallSettings};
 
 /**
  * A Promise implementation that supports deferred resolution.
@@ -138,44 +121,6 @@ export function isEmpty(value: {}): boolean {
  */
 export function isFunction(value: unknown): boolean {
   return typeof value === 'function';
-}
-
-/**
- * Determines whether the provided error is considered permanent for the given
- * RPC.
- *
- * @private
- */
-export function isPermanentRpcError(
-  err: GoogleError,
-  methodName: string
-): boolean {
-  if (err.code !== undefined) {
-    const retryCodes = getRetryCodes(methodName);
-    return retryCodes.indexOf(err.code) === -1;
-  } else {
-    return false;
-  }
-}
-
-/**
- * Returns the list of retryable error codes specified in the service
- * configuration.
- * @private
- */
-export function getRetryCodes(methodName: string): number[] {
-  return serviceConfig[methodName]?.retry?.retryCodes ?? [];
-}
-
-/**
- * Returns the backoff setting from the service configuration.
- * @private
- */
-export function getRetryParams(methodName: string): BackoffSettings {
-  return (
-    serviceConfig[methodName]?.retry?.backoffSettings ??
-    createDefaultBackoffSettings()
-  );
 }
 
 /**
